backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.1
supabase>=2.4.0
apscheduler>=3.10.4
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
openai>=1.30.0
pdfplumber>=0.11.0
playwright>=1.43.0
google-adk>=1.0.0
google-search-results>=2.4.2
praw>=7.7.0
tweepy>=4.14.0
